+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -156,13 +156,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case off
     }
 
-    enum PaymentMethodOptionsSetupFutureUsageEnabled: String, PickerEnum {
-        static var enumName: String { "PMO SFU" }
-
-        case on
-        case off
-    }
-
     struct PaymentMethodOptionsSetupFutureUsage: Codable, Equatable {
         // Supports all SFU values
         var card: SetupFutureUsageAll
@@ -586,11 +579,29 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case `continue`
     }
 
+    enum RowSelectionBehavior: String, PickerEnum {
+        static let enumName: String = "rowSelectionBehavior"
+        case `default`
+        case immediateAction
+    }
+
     enum CardBrandAcceptance: String, PickerEnum {
         static let enumName: String = "cardBrandAcceptance"
         case all
         case blockAmEx
         case allowVisa
+    }
+
+    enum LinkFlowControllerChanges: String, PickerEnum {
+        static let enumName: String = "Link FlowController RUX"
+        case on
+        case off
+    }
+
+    enum LinkDefaultOptIn: String, PickerEnum {
+        static let enumName: String = "Link DOI"
+        case on
+        case off
     }
 
     enum ConfigurationStyle: String, PickerEnum {
@@ -612,7 +623,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
     var merchantCountryCode: MerchantCountry
     var apmsEnabled: APMSEnabled
     var supportedPaymentMethods: String?
-    var paymentMethodOptionsSetupFutureUsageEnabled: PaymentMethodOptionsSetupFutureUsageEnabled
     var paymentMethodOptionsSetupFutureUsage: PaymentMethodOptionsSetupFutureUsage
 
     var shippingInfo: ShippingInfo
@@ -652,7 +662,10 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
     var collectAddress: BillingDetailsAddress
     var formSheetAction: FormSheetAction
     var embeddedViewDisplaysMandateText: DisplaysMandateTextEnabled
+    var rowSelectionBehavior: RowSelectionBehavior
     var cardBrandAcceptance: CardBrandAcceptance
+    var linkFlowControllerChanges: LinkFlowControllerChanges
+    var linkDefaultOptIn: LinkDefaultOptIn
 
     static func defaultValues() -> PaymentSheetTestPlaygroundSettings {
         return PaymentSheetTestPlaygroundSettings(
@@ -667,7 +680,6 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
             amount: ._5099,
             merchantCountryCode: .US,
             apmsEnabled: .on,
-            paymentMethodOptionsSetupFutureUsageEnabled: .off,
             paymentMethodOptionsSetupFutureUsage: PaymentMethodOptionsSetupFutureUsage.defaultValues(),
             shippingInfo: .off,
             applePayEnabled: .on,
@@ -705,7 +717,11 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
             collectAddress: .automatic,
             formSheetAction: .continue,
             embeddedViewDisplaysMandateText: .on,
-            cardBrandAcceptance: .all)
+            rowSelectionBehavior: .default,
+            cardBrandAcceptance: .all,
+            linkFlowControllerChanges: .off,
+            linkDefaultOptIn: .off
+        )
     }
 
     static let nsUserDefaultsKey = "PaymentSheetTestPlaygroundSettings"