+++ StripePaymentSheet/StripePaymentSheet/Source/Helpers/PaymentSheetLinkAccount.swift
@@ -15,7 +15,6 @@ protocol PaymentSheetLinkAccountInfoProtocol {
     var email: String { get }
     var redactedPhoneNumber: String? { get }
     var isRegistered: Bool { get }
-    var isLoggedIn: Bool { get }
 }
 
 struct LinkPMDisplayDetails {
@@ -49,6 +48,15 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
 
         // Clicked button in separate Link sheet
         case clicked_button_mobile_v1 = "clicked_button_mobile_v1"
+
+        // Checkbox pre-checked, w/ email & phone prefilled
+        case prechecked_opt_in_box_prefilled_all = "prechecked_opt_in_box_prefilled_all"
+
+        // Checkbox pre-checked, some fields prefilled
+        case prechecked_opt_in_box_prefilled_some = "prechecked_opt_in_box_prefilled_some"
+
+        // Checkbox pre-checked, no fields prefilled
+        case prechecked_opt_in_box_prefilled_none = "prechecked_opt_in_box_prefilled_none"
     }
 
     // Dependencies
@@ -73,10 +81,6 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
 
     var isRegistered: Bool {
         return currentSession != nil
-    }
-
-    var isLoggedIn: Bool {
-        return sessionState == .verified
     }
 
     var sessionState: SessionState {
@@ -283,6 +287,7 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
 
     func createPaymentDetails(
         linkedAccountId: String,
+        isDefault: Bool,
         completion: @escaping (Result<ConsumerPaymentDetails, Error>) -> Void
     ) {
         retryingOnAuthError(completion: completion) { completionRetryingOnAuthErrors in
@@ -295,8 +300,24 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
             session.createPaymentDetails(
                 linkedAccountId: linkedAccountId,
                 consumerAccountPublishableKey: self.publishableKey,
+                isDefault: isDefault,
                 completion: completionRetryingOnAuthErrors
             )
+        }
+    }
+
+    func listPaymentDetails(
+        supportedTypes: [ConsumerPaymentDetails.DetailsType]
+    ) async throws -> [ConsumerPaymentDetails] {
+        return try await withCheckedThrowingContinuation { continuation in
+            listPaymentDetails(supportedTypes: supportedTypes) { result in
+                switch result {
+                case .success(let details):
+                    continuation.resume(returning: details)
+                case .failure(let error):
+                    continuation.resume(throwing: error)
+                }
+            }
         }
     }
 
@@ -317,6 +338,33 @@ class PaymentSheetLinkAccount: PaymentSheetLinkAccountInfoProtocol {
                 consumerAccountPublishableKey: self.publishableKey,
                 completion: completionRetryingOnAuthErrors
             )
+        }
+    }
+
+    func listShippingAddress() async throws -> ShippingAddressesResponse {
+        return try await withCheckedThrowingContinuation { continuation in
+            listShippingAddress { result in
+                switch result {
+                case .success(let response):
+                    continuation.resume(returning: response)
+                case .failure(let error):
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    func listShippingAddress(
+        completion: @escaping (Result<ShippingAddressesResponse, Error>) -> Void
+    ) {
+        retryingOnAuthError(completion: completion) { completionRetryingOnAuthErrors in
+            guard let session = self.currentSession else {
+                stpAssertionFailure()
+                completion(.failure(PaymentSheetError.unknown(debugDescription: "Paying with Link without valid session")))
+                return
+            }
+
+            session.listShippingAddress(with: self.apiClient, consumerAccountPublishableKey: self.publishableKey, completion: completionRetryingOnAuthErrors)
         }
     }
 
@@ -503,7 +551,8 @@ extension PaymentSheetLinkAccount {
     func makePaymentMethodParams(
         from paymentDetails: ConsumerPaymentDetails,
         cvc: String?,
-        billingPhoneNumber: String?
+        billingPhoneNumber: String?,
+        allowRedisplay: STPPaymentMethodAllowRedisplay?
     ) -> STPPaymentMethodParams? {
         guard let currentSession = currentSession else {
             stpAssertionFailure("Cannot make payment method params without an active session.")
@@ -511,6 +560,9 @@ extension PaymentSheetLinkAccount {
         }
 
         let params = STPPaymentMethodParams(type: .link)
+        if let allowRedisplay {
+            params.allowRedisplay = allowRedisplay
+        }
         params.billingDetails = STPPaymentMethodBillingDetails(billingAddress: paymentDetails.billingAddress, email: paymentDetails.billingEmailAddress)
         params.billingDetails?.phone = billingPhoneNumber
         params.link?.paymentDetailsID = paymentDetails.stripeID