+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/Link/Controllers/PayWithLinkViewController-NewPaymentViewController.swift
@@ -141,7 +141,7 @@ extension PayWithLinkViewController {
             super.viewDidLoad()
             addChild(addPaymentMethodVC)
 
-            view.backgroundColor = .linkBackground
+            view.backgroundColor = .linkSurfacePrimary
 
             addPaymentMethodVC.view.backgroundColor = .clear
             errorLabel.isHidden = true
@@ -160,21 +160,9 @@ extension PayWithLinkViewController {
             stackView.setCustomSpacing(LinkUI.extraLargeContentSpacing, after: addPaymentMethodVC.view)
             stackView.translatesAutoresizingMaskIntoConstraints = false
 
-            let scrollView = LinkKeyboardAvoidingScrollView()
-            #if !os(visionOS)
-            scrollView.keyboardDismissMode = .interactive
-            #endif
-            scrollView.addSubview(stackView)
-
-            contentView.addAndPinSubview(scrollView)
+            contentView.addAndPinSubviewToSafeArea(stackView)
 
             NSLayoutConstraint.activate([
-                stackView.topAnchor.constraint(equalTo: scrollView.topAnchor, constant: preferredContentMargins.top),
-                stackView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor, constant: -preferredContentMargins.bottom),
-                stackView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
-                stackView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor),
-                stackView.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
-
                 titleLabel.leadingAnchor.constraint(
                     equalTo: stackView.safeAreaLayoutGuide.leadingAnchor,
                     constant: preferredContentMargins.leading),
@@ -241,6 +229,11 @@ extension PayWithLinkViewController {
                         // part of the payment details' billing information.
                         billingPhoneNumber: confirmParams.paymentMethodParams.billingDetails?.phone
                     )
+
+                    guard !context.launchedFromFlowController else {
+                        coordinator?.handlePaymentDetailsSelected(paymentDetails, confirmationExtras: confirmationExtras)
+                        return
+                    }
 
                     self.coordinator?.confirm(
                         with: self.linkAccount,
@@ -314,14 +307,12 @@ extension PayWithLinkViewController {
         @objc
         func cancelButtonTapped(_ sender: Button) {
             if isAddingFirstPaymentMethod {
-                coordinator?.cancel()
+                coordinator?.cancel(shouldReturnToPaymentSheet: false)
             } else {
-                navigationController?.popViewController(animated: true)
+                _ = bottomSheetController?.popContentViewController()
             }
         }
-
     }
-
 }
 
 extension PayWithLinkViewController.NewPaymentViewController: AddPaymentMethodViewControllerDelegate {