+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## XX.XX.X XXXX-XX-XX
+
+### Connect
+* [CHANGED] Removed `PrivateBetaConnect` annotation and all usages. [Connect SDK](https://docs.stripe.com/connect/get-started-connect-embedded-components?platform=ios) is now GA.
+
+### PaymentSheet
+* [Added] The ability to customize form insets with `PaymentSheetAppearance.formInsets`.
+
+## 24.14.0 2025-05-19
+### PaymentSheet
+* [Added] Support for payment method options setup future usage in public preview.
+
 ## 24.13.1 2025-05-05
 ### PaymentSheet
 * [Fixed] Updated the "TEST MODE" badge to say "TEST", due to the upcoming release of [Sandboxes](https://docs.stripe.com/sandboxes).